python-requirements.txt
xmltodict
numpy
networkx == 1.11
biopython == 1.72
freesasa >= 2.0.3
scikit-learn >= 0.18.1
